eth_listener/requirements.txt
web3==6.15.1
python-osc==1.8.3
python-dotenv==1.0.1
